infra/lambda/requirements.txt
psycopg2-binary==2.9.10
google-cloud-bigquery==3.27.0
google-auth==2.37.0
